adds try wrapper to testgrid cloud-init script

+++ testgrid/deploy/tg-script.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+# try wraps commands that might fail due to external internet factors
+function try () {
+    ${1} || {
+        echo "FAILED: ${1}"
+        exit 1
+    }
+}
+
 apt update
 DEBIAN_FRONTEND=noninteractive apt upgrade -y
 
@@ -12,7 +20,7 @@ mount /dev/nvme0n1 /var/openebs/local/
 echo "Installing kURL"
 INSTALL_SCRIPT=/root/kurl-install.sh
 curl https://kurl.sh/f3dd2d4 > $INSTALL_SCRIPT
-sed -i 's/parse_yaml_into_bash_variables$/parse_yaml_into_bash_variables\n    PRIVATE_ADDRESS=$(\/sbin\/ifconfig bond0:0 | awk "\/inet \/ {print \\$2}")/' $INSTALL_SCRIPT 
+sed -i 's/parse_yaml_into_bash_variables$/parse_yaml_into_bash_variables\n    PRIVATE_ADDRESS=$(\/sbin\/ifconfig bond0:0 | awk "\/inet \/ {print \\$2}")/' $INSTALL_SCRIPT
 chmod +x $INSTALL_SCRIPT
 ./$INSTALL_SCRIPT
 [ -d /root/.kube ] || mkdir /root/.kube
@@ -24,16 +32,17 @@ kubectl delete deployment openebs-admission-server -n openebs
 
 echo "Instaling KubeVirt"
 kubectl create namespace kubevirt
-kubectl apply -f https://github.com/kubevirt/kubevirt/releases/download/v0.32.0/kubevirt-operator.yaml
-kubectl apply -f https://github.com/kubevirt/kubevirt/releases/download/v0.32.0/kubevirt-cr.yaml
+try "kubectl apply -f https://github.com/kubevirt/kubevirt/releases/download/v0.32.0/kubevirt-operator.yaml"
+try "kubectl apply -f https://github.com/kubevirt/kubevirt/releases/download/v0.32.0/kubevirt-cr.yaml"
 kubectl wait --timeout=180s --for=condition=Available -n kubevirt kv/kubevirt
 kubectl create ns cdi
-kubectl apply -f https://github.com/kubevirt/containerized-data-importer/releases/download/v1.24.0/cdi-operator.yaml
-kubectl apply -f https://github.com/kubevirt/containerized-data-importer/releases/download/v1.24.0/cdi-cr.yaml
+try "kubectl apply -f https://github.com/kubevirt/containerized-data-importer/releases/download/v1.24.0/cdi-operator.yaml"
+try "kubectl apply -f https://github.com/kubevirt/containerized-data-importer/releases/download/v1.24.0/cdi-cr.yaml"
 
 echo "Installing krew KubeVirt plugin"
 export HOME=/root
 curl https://krew.sh/virt | bash
+[ -f /root/.krew/bin/kubectl-virt ] || exit 1
 
 echo "Setting up tgrun service"
 cat <<-TGRUND > /lib/systemd/system/tgrun.service
@@ -62,6 +71,7 @@ ExecStart=/bin/bash -c '/bin/tgrun run'
 [Install]
 WantedBy=multi-user.target
 TGRUND
+systemctl daemon-reload
 
 echo "pulling tgrun image and extracting binary"
 docker pull replicated/tgrun:latest
